underline rule - tweak it a bit
Summary:
basically what you told me to do in D7995, but I also added a "/" to the 'don't capture' clause. In my chosen test code without this check we would end up underlining "and also", and I figured this special case will probably need to handle multiple files in a row with this formatting.

Of course, this means __/something__ also won't be underlined.

Test Plan: plugged the regex into a handy online tool (http://regex101.com/) and played with it a bunch. re-ran unit tests and they passed

Reviewers: epriestley

Reviewed By: epriestley

CC: Korvin, epriestley, aran

Differential Revision: https://secure.phabricator.com/D7999