Attempt to automatically build XHPAST.
Summary: Fixes T4951. Whilst `xhpast` was originally intended to be an internal development tool, it is not unreasonable for `ArcanistXHPASTLinter` to be used externally. Currently, however, if the compiled version of `xhpast` is not up-to-date, we simply exit and ask the user to build XHPAST themselves. Instead, let's try to build XHPAST automatically.

Test Plan: Removed `./src/parser/xhpast/bin/xhpast` and ran `arc lint` over a PHP file. Verified that the `xhpast` binary was created during this process.

Reviewers: epriestley, #blessed_reviewers

Reviewed By: epriestley, #blessed_reviewers

Subscribers: epriestley, Korvin

Maniphest Tasks: T4951

Differential Revision: https://secure.phabricator.com/D9076