Ensure empty lines do not appear with large widths on last column of PhutilConsoleTable
Summary:
If the last column has a large width, it will space out the content of the last column even if no borders are shown.  This causes empty lines to appear (that have no content).

Instead, no spacing should be performed on the data of the last column if the column is left aligned and no borders are shown.  Long data in the last column will wrap in the terminal (as it would anyway), but rows with short data won't cause empty lines.

Test Plan: Tested under Omni.

Reviewers: #blessed_reviewers, epriestley

Reviewed By: #blessed_reviewers, epriestley

Subscribers: Korvin, epriestley

Differential Revision: https://secure.phabricator.com/D11328