Attempt to build XHPAST during unit tests
Summary: If we have an out-of-date XHPAST binary, attempt to rebuild XHPAST as a part of running the unit tests. If XHPAST is missing (or we cannot build it for whatever reason), just skip the unit tests as per the existing behavior.

Test Plan: Deleted `src/parser/xhpast/bin/xhpast` and ran `arc unit -- src/parser/xhpast`.

Reviewers: epriestley, #blessed_reviewers

Reviewed By: epriestley, #blessed_reviewers

Subscribers: Korvin, epriestley

Differential Revision: https://secure.phabricator.com/D11516