Allow callers to explicitly request a progress bar redraw
Summary:
Ref T7149. Normally, we only redraw the bar every 100ms.

When resuming large file uploads, we "complete" the first part of the file immediately, but it takes much less than 100ms, and then we may sit for the first chunk without updating. This is a little confusing. Allow a forced redraw so we can sit at "Resuming: 60%" or whatever instead.

Test Plan: See next diff.

Reviewers: btrahan

Reviewed By: btrahan

Subscribers: epriestley

Maniphest Tasks: T7149

Differential Revision: https://secure.phabricator.com/D12081