Unescape sequences in .cow files which look like escaped sequences
Summary:
Cowsay's .cow files are perl scripts which contain escaped character sequences. This results in an excess of backslashes in the rendered cow. This change replaces occurrences of `\x` with just `x`.

This change also introduces conditional unescaping based on whether the cowfile is an original perl-script cowfile or a plain ascii-art file.

Refs T9471

Test Plan:
Use cowsay in a comment and verify that the rendered cow displays appropriately.
Use cowsay with the "dragon" cow and verify that the rendered dragon displays appropriately (contains `\@` sequence).

Reviewers: #blessed_reviewers, epriestley

Reviewed By: #blessed_reviewers, epriestley

Subscribers: Korvin

Maniphest Tasks: T9471

Differential Revision: https://secure.phabricator.com/D14170