Use new format for docblock specials when parsing symbols
Summary: Fixes T11575. Before, specials used to be separated by newlines, now they are returned as an array for us, //or// returned as a single value for us.  Casting the result to an array is the minimum amount of disruption we can incur here to keep the existing logic in place and provide the same values.

Test Plan:
 - deleted my `src/.phutil_module_cache` from arcanist, libphutil and phabricator
 - ran `arc liberate src` in all of those places
 - ran all of their unit tests
 - saw all tests pass and no modifications made to `__phutil_library_map__.php` for any of the libraries

Reviewers: #blessed_reviewers, epriestley, joshuaspence

Reviewed By: #blessed_reviewers, epriestley, joshuaspence

Subscribers: epriestley

Maniphest Tasks: T11575

Differential Revision: https://secure.phabricator.com/D16484