Make "fail quietly" flag in DeferredLog a little more quiet
Summary:
Ref T12611. Currently, access logs have a "setFailQuietly(true)" flag set on them (rationale: the webserver should still work if the log volume fills up or logs are otherwise not correctly configured) but it's only somewhat quiet.

Make it quieter when the log directory itself (not just the logfile) doesn't exist or can't be created.

(If installs run into confusion with this, we could more reasonably raise a setup warning instead.)

Test Plan: With a missing logfile directory, saw a log about being unable to log instead of an exception.

Reviewers: chad, amckinley

Reviewed By: amckinley

Maniphest Tasks: T12611

Differential Revision: https://secure.phabricator.com/D17777