Make empty Remarkup table rows have more intuitive behavior
Summary:
Fixes T12849. Currently, when a row contains only empty cells and cells with "----" in them, we treat it as indicating which cells above should be rendered with `<th />` instead of `<td />`.

This produces an unintuitive result when //every// cell is empty.

Instead, treat a row as a "headings" row only if it has at least one "---" cell in it.

Test Plan:
Added a unit test.

Note that the resulting row is empty, so it doesn't get normal line height.

{F5008070}

Reviewers: chad

Reviewed By: chad

Maniphest Tasks: T12849

Differential Revision: https://secure.phabricator.com/D18140