Fix a typo in JSON parser
Summary:
Caught this in error logs. This is an external and it looks like this is already fixed in the upstream, but updating just for a typo fix feels like a bit much.

(<https://github.com/Seldaek/jsonlint/commit/bc020395df50bb658bf468bad8bfc962275c4a2a>)

Test Plan: Read it.

Reviewers: amckinley

Reviewed By: amckinley

Differential Revision: https://secure.phabricator.com/D18752