Make FileFinder's withName() + withSuffix() work like every other Query class
Summary:
See PHI336. See <https://discourse.phabricator-community.org/t/using-the-filefinder-class-withsuffix-and-withname-return-entries-with-either-the-suffix-or-the-name-match/1035?u=dangerouslyil>.

Currently, `withName()` and `withSuffix()` mean "with either the name or the suffix", which isn't how any other `Query` class works.

Note that "name" means "exact name", per `find` semantics, so this combination isn't necesarily terribly useful. However, the current behavior is clearly surprising given that everything else in the project works differently.

Test Plan: Added a test case for both constraints, made it pass.

Reviewers: amckinley

Reviewed By: amckinley

Subscribers: yelirekim

Differential Revision: https://secure.phabricator.com/D18979