Fix Filesytem::walkToRoot() on Windows with default ("/") root
Summary:
Fixes T9459. See PHI817. An install is reporting that `arc upgrade` fails on Windows, failing to recognize that `arcanist/` is a Git working copy.

This appears to reproduce for me, and boil down to `FileList->contains()` testing if `/` contains `C:\Application Data32\Roaming\My Documents\Users\epriestley\devtools\arcanist`. It (literally) does not, but our intent is that it should.

Instead of defaulting the root to `/`, treat it somewhat more specially so `Filesystem::walkToRoot(X)` on Windows always walks all the way up.

Test Plan: This is slightly shaky since my Windows "development environment" is a mess, but I was able to run `arc upgrade` with this patch applied and make it through working copy recognition. I'll re-test after landing.

Reviewers: amckinley

Reviewed By: amckinley

Maniphest Tasks: T9459

Differential Revision: https://secure.phabricator.com/D19591